+++ kernel/app/src/js/config.js
@@ -49,6 +49,52 @@ window.publish.setFontSize = (fontSize, fontFamily, codeLigatures) => {
     return null;
 }
 
+/**
+ * 创建图标
+ * @params {string} icon: 图标名
+ * @params {string} className: 类名
+ * @params {string} id: 块 ID
+ * @returns {HTMLElement} 图标
+ */
+window.publish.createIcon = (icon, className, id = null) => {
+    const element = document.createElement('div');
+    element.innerHTML = `
+        <svg class="${className}">
+            <use xlink:href="${icon}"></use>
+        </svg>
+    `;
+    if (id) element.firstElementChild.dataset.nodeId = id;
+    return element.firstElementChild;
+}
+
+/**
+ * 创建面包屑项
+ * @params {string} text: 面包屑项的文本
+ * @params {string} id: 面包屑项的 ID
+ * @params {string} icon: 面包屑项的图标
+ * @params {bool} active: 面包屑项是否激活
+ * @returns {HTMLElement} 面包屑项
+ */
+window.publish.createBreadcrumbItem = (text, id = null, icon = '#iconFile', active = false) => {
+    const breadcrumb__item = document.createElement('span');
+    breadcrumb__item.classList.add('protyle-breadcrumb__item');
+    if (active) breadcrumb__item.classList.add('protyle-breadcrumb__item--active');
+    if (id) breadcrumb__item.dataset.nodeId = id;
+
+    const breadcrumb__icon = window.publish.createIcon(icon, 'popover__block', id);
+    const breadcrumb__text = document.createElement('span');
+    breadcrumb__text.classList.add(
+        'protyle-breadcrumb__text',
+        'protyle-breadcrumb__text--ellipsis',
+    );
+    breadcrumb__text.innerText = text;
+    breadcrumb__item.title = text;
+
+    breadcrumb__item.appendChild(breadcrumb__icon);
+    breadcrumb__item.appendChild(breadcrumb__text);
+    return breadcrumb__item;
+}
+
 (() => {
     const url = new URL(window.location.href);
 
@@ -130,5 +176,25 @@ window.publish.setFontSize = (fontSize, fontFamily, codeLigatures) => {
                     }
                     break;
             }
+    }
+
+    /* 添加面包屑 */
+    const breadcrumb = document.getElementById('breadcrumb');
+    if (breadcrumb) {
+        paths = window.publish.document.path.split('/');
+        hpaths = window.publish.document.hpath.split('/');
+        const depth = Math.min(paths.length, hpaths.length) - 1; // 当前文档的深度
+        for (let i = 0; i <= depth; i++) {
+            breadcrumb.appendChild(window.publish.createIcon(
+                '#iconRight',
+                'protyle-breadcrumb__arrow',
+            ));
+            breadcrumb.appendChild(window.publish.createBreadcrumbItem(
+                hpaths[i],
+                paths[i],
+                '#iconFile',
+                i === depth,
+            ));
+        }
     }
 })()