+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 更改日志 | CHANGE LOG
 
 - 提高块跳转超链接与子窗口的 `z-index` 层级 | Raise the 'z-index' level of block jump hyperlinks and popover windows.
+- 将子窗口预览功能默认触发时间调整为 `1000ms` | Adjust the time of the default trigger time of the popover preview window to `1000ms`.
 
 ## v0.1.1 / 2022-07-25