+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 - 提高块跳转超链接与子窗口的 `z-index` 层级 | Raise the 'z-index' level of block jump hyperlinks and popover windows.
 - 将子窗口预览功能默认触发时间调整为 `1000ms` | Adjust the time of the default trigger time of the popover preview window to `1000ms`.
 - 使用 `getDoc` API 查询文档的 DOM | Use the `getDoc` API to query the DOM of the document.
+- 为文档添加面包屑 | Add a breadcrumb for the document.
 
-## v0.1.1 / 2022-07-25
+## v0.1.1 / 2022-07-28
 
 - [v0.1.0 ... v0.1.1](https:///github.com/Zuoqiu-Yingyi/siyuan-publish/compare/v0.1.0...v1.1.1)
 - 为每个块添加跳转超链接 | Add a link to each block to jump to it.